Q-BRIDGE: Handle missing dot1qTpFdbAddress column
We've seen data that indicates some Q-BRIDGE agent implementations don't
return the dot1qTpFdbAddress column because it's part of the table's
index, and can therefore be parsed. This fix first checks for the
column's value, but parses the MAC address from the table's index if it
can't be found.

Fixes ZEN-25336.