Rewrite Redis backend to fix inconsistencies
This is a big update that replaces the "catalog" Redis backend with a
redesigned Redis backend. The existing Redis backend has a Zope
catalog-like API in front of it because the original backend used a Zope
catalog, and when that was found not to be suitable, it was found to be
less invasive to keep the catalog-like API in front of a Redis backend.
This update does away with the catalog-like API since it isn't
necessary, and the change was going to be invasive anyway.

The reason for replacing the backend is to deal with a problem of one
connections provider overwriting connection information from other
connections providers. The design had no concept or attribution or
origination for the data. The new design has origination at its heart.
Each connection provider is identified as a "origin", and every bit
connection data it contributes is tagged to that origin. Only when a bit
of data has no origin tags left can be be removed.

The "connections_catalog" module has been entirely removed. It has been
replaced by the "connections" module. The "connections" module is a thin
API used by the rest of the ZenPack to read and write connections data.
Most of the work is done by the "graph" module which is only used by the
"connections" module.

Fixes ZEN-25720, ZEN-25724, ZEN-25950.