Handle objects deleted from ZODB but not Redis
This is required because the Redis database backing the network map is
not consistent with ZODB. Specifically, devices and component entries in
the Redis database don't immediately get cleaned up when the devices and
components are deleted from ZODB.

The zenmapper service ultimately cleans these up on its next run, but we
have to work around stale entries regularly.

Fixes ZPS-88, ZPS-90.