Fix POSKeyError on first model of device
I didn't think a ZenPack could cause a POSKeyError, but here's a
completely reproducible way to create them. If you call _p_invalidate()
on a new object where _p_jar is is still None instead of a
ConnectionManager, you will get a POSKeyError when the transaction
attempts to commit.

Fixes ZPS-321.