Only consider L2 paths when pruning redundant MACs
The problem in ZPS-322 was that I classified nodes along one shortest
path from endpoint nodes to endpoint nodes as being "useful". Any MAC
nodes that weren't useful were pruned. This resulted in way too many MAC
nodes getting pruned because there was often a much shorter L3 path.

This fix is two-fold.

1. Only consider layer2 nodes for the shortest path calculation.
2. Unfavorably weight MAC node without a path.

This results in the right amount of MAC address nodes being left, and it
results in MAC address nodes without a path being much more likely to be
pruned. This is good, because the whole point of the "Show MAC
addresses" option on the network map is to allow the user to click on
MAC addresses to navigate directly to their associated interface.

There are some other small unrelated changes.

- Changed the node limit from 2,000 to 1,000. Improved error message.
- Documented new 1,000 node limit.
- Documented that the layers filters are "AND", not "OR".

Fixes ZPS-322.