Fix modeling performance with many component types
Switch to updating Layer2 information in
ModelerService.remote_applyDataMaps() instead of Device.setLastChange().
It turns out that for devices that submit many DataMaps while modeling
such as vSphere, the Device.setLastChange() method is called once for
every DataMap that results in a change. We only need to be called once
for the device if anything on it changed.

As an example, our large vcsim1 vCenter simulator results in over 24,000
DataMaps being applied for its initial model. This would result in the
Layer2 information in Redis being updated up to that many times when it
only had to happen once. So we're talking about a 24,000 times speedup
in that extreme case.

Fixes ZPS-736.