Fix infinite loop in event suppression logic
There was a bug in Suppressor.get_shortest_paths that could result in an
infinite loop. Fixing the bug by properly implementing the shortest path
algorithm that was in place resulted in the algorithm taking too far too
long to execute on large topologies to be feasible during event
processing.

Newer versions of the networkx library such as version 11 shipped with
Zenoss 5 have a shortest_simple_paths algorithm that's optimal for this
use case. However, this algorithm doesn't exist in the networkx 1.3
version shipped with Zenoss 4.2.5. So I've copied the BSD-licensed code
into this ZenPack for use when it isn't available in the installed
networkx library.

Using the new algorithm the functional tests all pass, and the manual
performance tests show no sign of a performance regression. If anything
there is a performance improvement. See the following Google spreadsheet
for the results of this performance testing. The "5.1.x" sheet is before
this change, and the "5.3.x" sheet is after. Note that 5.3.x has a
higher event processing rate overall. This is likely just a difference
in the systems I ran the tests on. The relative performance vs. baseline
is similar.

https://docs.google.com/a/zenoss.com/spreadsheets/d/1r5g6IzqvX613Q6C9vjokxrQyQYqsXawuSlvOGOcXkno/edit?usp=sharing

Fixes ZPS-1353.