Update zenmapper service and nest under /Zenoss/
Fresh installs of Layer2 1.4.0 will put the zenmapper service in the
/Zenoss/ organizer instead of at the root level as in previous versions
of Layer2. This change adds a migration to remove existing zenmapper
services at the root level, and install the most recent zenmapper
service configuration in the /Zenoss/ organizer. This will cause
upgrades to match fresh installations.

This also causes a number of updates to be made to the zenmapper
service.

- Moves zenmapper service from / to /Zenoss/. (ZPS-1352)
- Adds "debug" and "stats" actions.
- Adds now-standard "--logfileonly" option to command.
- Adds "clear", "update", and "force" service commands.
- Adds global.conf configuration file. (ZPS-2216)
- Adds controlplane_consumer endpoint so daemon stats can be published.
- Restricts instances configuration to 1 only. (ZPS-2144)
- Sets RAMCommitment to 2G in light of recent memory savings.