Fix "Set the root device or component" flares
There was some confusion about how to handle hashes between the two
different scopes in with the network map is displayed: global and
device. I made the functions that needed to be aware of scope aware of
it, and made the synchronization of the sidebar settings and the URL
query string values more explicit.

Fixes ZPS-1287 & ZPS-2463.