Remove DISTINCT to improve MySQL performance
It was found that the DISTINCT qualifier on these queries is unnecessary
because the merge_layers method is performing a distinct operation of
its own. Furthermore it was found that during large-scale parallel
modeling with Impact installed that the DISTINCT qualifier doubled the
CPU usage of mariadb-model due to creation of temporary tables.

Fixes ZPS-2550.