Fix ZODB growth caused by zenmapper
The primary culprit was broken change detection. Each cycle zenmapper
was only supposed to process nodes (devices) that had changed since
their l2_edges information was last written to the database. Due to the
underlying behavior of "INSERT IGNORE" that was being used, the last
change recorded in the l2_providers was never being updated. This means
that once a device changed a second time, it would have to be processed
on every subsequent cycle of zenmapper. That problem has been fixed.

Even with that, zenmapper will cause database fragmentation. So I also
added a periodic (daily) optimization of all of the l2_* tables.

Fixes ZPS-3548.