Optimize MySQL usage to avoid table fragmentation
The previous MySQL usage used a strategy of tracking all edges by the
"provider" that created them. A provider is an IConnectionsProvider
adapter. Anytime a change was detected in a provider either during
modeling or by zenmapper, all of the previous edges from that provider
were deleted from the l2_edges table, and the new set of edges were
inserted. This resulted in a lot of fragmentation for two reasons. We
were deleting a lot of rows, and the rows contained variable-length
columns so the space vacated by the deleted rows couldn't be guaranteed
to be reusable by the new rows. The problem would be exacerbated if
MySQL's purge thread fell behind.

The new strategy in this commit keeps the provider concept intact, but
instead of deleting and replacing all of the provider's edges when any
change is detected, a "diff" is performed so that only old edges that no
longer exist are deleted, and only new edges are inserted. Additionally
the new l2_v2_edges table only has fixed-length columns so that even the
fewer rows that are deleted can be more easily reused by new rows.

Previously the following tables were used in the zodb (mariadb-model)
database.

- l2_metadata
- l2_providers
- l2_edges

Now we have the following more normalized tables.

- l2_v2_metadata
- l2_v2_providers
- l2_v2_layers
- l2_v2_nodes
- l2_v2_edges

In testing this results in faster operation, a smaller minimal size on
disk, and drastically reduced table fragmentation.

This change also adds a new "optimize-interval" configuration option to
the zenmapper service. This option allows the user to configure how
frequently all of the aforementioned database tables should be
optimized. Previously this would happen once every 24 hours and was not
configurable. The new default for this option is 0 which means that
optimization will never happen automatically.

Code was also added to zenmapper to drop the l2_metadata, l2_providers,
and l2_edges tables that are no longer used.

Fixes ZPS-4402.